latest refactor dry principle
