+++ src/services/controllers/user/actions/create.ts
@@ -3,8 +3,8 @@ import { UserModel } from "@src/models/user";
 import mH from "@src/messages/constants/http";
 import mU from "@src/messages/constants/user";
 import mO from "@src/messages/constants/others";
-import { findEntity } from "@src/utilities/misc";
-import { returnCheckMessage } from "@src/utilities/misc";
+import { findEntity } from "@src/utilities/misc/find";
+import { returnCheckMessage } from "@src/utilities/misc/check";
 import { encryptPassword } from "@src/utilities/password";
 import { generateOTPAndExpiry } from "@src/utilities/otp";
 import { createObject } from "@src/utilities/crudFactory/create";
