+++ src/utilities/misc/check/index.ts
@@ -1,0 +1,7 @@
+export const checkJSONBody = (bodyData: object) => {
+  return JSON.parse(JSON.stringify(bodyData));
+};
+
+export const returnCheckMessage = (message: string) => {
+  throw message;
+};
