+++ src/utilities/misc/find/index.ts
@@ -1,13 +1,5 @@
 import mongoose from "mongoose";
-import { ModelObject, User, Product } from "@src/types";
-
-export const checkJSONBody = (bodyData: object) => {
-  return JSON.parse(JSON.stringify(bodyData));
-};
-
-export const returnCheckMessage = (message: string) => {
-  throw message;
-};
+import { ModelObject } from "@src/types";
 
 export const findEntity: Function = async (
   objectModel: mongoose.Model<object>,
@@ -36,19 +28,6 @@ export const findEntityAndUpdateFields: Function = async (
     );
 
     return { message: objectFieldsToUpdate };
-  } catch (error: unknown) {
-    throw `${error}`;
-  }
-};
-
-export const newInputValidationData: Function = async (
-  infoData: User | Product,
-  joiValidationFn: Function,
-): Promise<User | Error> => {
-  try {
-    const validateData = joiValidationFn(infoData);
-
-    return await validateData();
   } catch (error: unknown) {
     throw `${error}`;
   }
