+++ src/utilities/otp/index.ts
@@ -1,7 +1,7 @@
 import { multiply, sum } from "ramda";
 import mC from "@src/messages/constants/otp";
 import { setUserResendCodeToTrue } from "@src/utilities/user";
-import { returnCheckMessage } from "@src/utilities/misc";
+import { returnCheckMessage } from "@src/utilities/misc/check";
 
 export const generateOTPAndExpiry = () => {
   const dateExpiration = new Date();
