+++ src/validations/product/index.ts
@@ -1,22 +1,9 @@
 import Joi from "joi";
 import { Product } from "@src/types";
-import mV from "@src/messages/constants/validation";
+import { validateProductBody } from "@src/validations/helper/productValidation";
 
 export const productValidation = function (productBody: Product) {
   const productBodyForChecking = Object.assign({}, Object.freeze(productBody));
-
-  function validateProductBody() {
-    return Joi.object({
-      category: Joi.string().min(mV.min_string).max(mV.max_string).required(),
-      name: Joi.string().min(mV.min_string).max(mV.max_string).required(),
-      description: Joi.string()
-        .min(mV.min_string)
-        .max(mV.max_string)
-        .required(),
-      price: Joi.number().min(mV.min_price).max(mV.max_price).required(),
-      qty: Joi.number().min(mV.min_qty).max(mV.max_qty).required(),
-    });
-  }
 
   return async function () {
     const { category, name, description, price, qty } = productBodyForChecking;
