+++ src/validations/user/index.ts
@@ -1,33 +1,11 @@
-import Joi from "joi";
 import { User } from "@src/types";
-import mO from "@src/messages/constants/others";
-import mV from "@src/messages/constants/validation";
+import { validateUserBodyData } from "@src/validations/helper/userValidation";
 
 export const userValidation = function (userBodyData: User) {
   const userBodyDataForChecking = Object.assign(
     {},
     Object.freeze(userBodyData),
   );
-
-  function validateUserBodyData() {
-    return Joi.object({
-      name: Joi.string().min(mV.min_string).max(mV.max_string).required(),
-      address: Joi.string().min(mV.min_string).max(mV.max_string).optional(),
-      dob: Joi.date().optional(),
-      email: Joi.string()
-        .email({ tlds: { allow: mO.no } })
-        .required(),
-      cellNumber: Joi.string()
-        .min(mV.min_cell_num)
-        .max(mV.max_cell_num)
-        .optional(),
-      gender: Joi.string().min(mV.min_gender).max(mV.max_gender).optional(),
-      password: Joi.string()
-        .min(mV.min_password)
-        .max(mV.max_password)
-        .optional(),
-    });
-  }
 
   return async function () {
     const { name, address, dob, email, cellNumber, gender, password } =
